+++ src/util/EntityToFormMapper.js
@@ -22,7 +22,8 @@ const DEFAULTS = {
     falseLabel: 'False',
     nillLabel: 'N/A'
   },
-  showNillableBooleanOption: true
+  showNillableBooleanOption: true,
+  showNonVisibleAttributes: false
 }
 
 // Create an object type UserException
@@ -214,14 +215,16 @@ const getHtmlFieldType = (fieldType: EntityFieldType): HtmlFieldType => {
 
 /**
  * If there is a visible expression present, return a function which evaluates the expression.
- * If there is no expression present, return a function which evaluates to the value of attribute.visible
+ * If there is no expression present, check if mapper is run with showVisibleAttribute option set to true,
+ * if this is not the case attributes visible property is used
  *
  * @param attribute
+ * @param mapperOptions
  * @returns {Function} Function which evaluates to a boolean
  */
-const isVisible = (attribute): ((?Object) => boolean) => {
+const isVisible = (attribute, mapperOptions: MapperSettings): ((?Object) => boolean) => {
   const expression = attribute.visibleExpression
-  return expression ? (data) => evaluator(expression, data) : () => attribute.visible
+  return expression ? (data) => evaluator(expression, data) : () => mapperOptions.showNonVisibleAttributes || attribute.visible
 }
 
 /**
@@ -337,7 +340,7 @@ const generateFormSchemaField = (attribute, entityMetadata:any, mapperOptions: M
     required: isRequired(attribute),
     disabled: isDisabled,
     readOnly: isDisabled,
-    visible: isVisible(attribute),
+    visible: isVisible(attribute, mapperOptions),
     validate: isValid(attribute),
     unique: buildIsUniqueFunction(attribute, entityMetadata, mapperOptions)
   }
@@ -447,6 +450,7 @@ const buildMapperSettings = (settings?: MapperOptions): MapperSettings => {
   }
 
   const mapperMode = settings.mapperMode ? settings.mapperMode : DEFAULTS.mapperMode
+
   let booleanLabels = DEFAULTS.booleanLabels
   if (settings.booleanLabels) {
     booleanLabels = {
@@ -455,15 +459,22 @@ const buildMapperSettings = (settings?: MapperOptions): MapperSettings => {
       nillLabel: settings.booleanLabels.nillLabel ? settings.booleanLabels.nillLabel : 'N/A'
     }
   }
+
   let showNillableBooleanOption = DEFAULTS.showNillableBooleanOption
   if (typeof (settings.showNillableBooleanOption) === 'boolean') {
     showNillableBooleanOption = settings.showNillableBooleanOption
   }
 
+  let showNonVisibleAttributes = DEFAULTS.showNonVisibleAttributes
+  if (typeof (settings.showNonVisibleAttributes) === 'boolean') {
+    showNonVisibleAttributes = settings.showNonVisibleAttributes
+  }
+
   return {
-    mapperMode: mapperMode,
-    booleanLabels: booleanLabels,
-    showNillableBooleanOption: showNillableBooleanOption
+    mapperMode,
+    booleanLabels,
+    showNillableBooleanOption,
+    showNonVisibleAttributes
   }
 }